refactor(event-log): deduplicate hash and test helper functions (closes #79)
Production code:
- Make hash_pair, compute_event_canonical_hash, and event_type_tag
  pub(crate) in tree.rs so proof.rs and other modules can import them
- Remove duplicate hash_pair from proof.rs, replaced with import

Test code:
- Create event_log/test_helpers.rs with shared test_keypair,
  did_from_pubkey (standardized to return DID), sign_event,
  leaf_hash_from_event, and build_test_log
- Remove duplicate helpers from tree.rs, proof.rs, checkpoint.rs,
  and metrics.rs test modules (replaced with imports)
- Update bridge/claiming.rs to import shared test_keypair and
  did_from_pubkey instead of maintaining local copies

Integration test (phase2_integration.rs) retains its own copies of
compute_event_canonical_hash and event_type_tag because tests/ cannot
access pub(crate) or #[cfg(test)] items. Copies are documented with
comments referencing issue #79.

Net: -431 lines, +42 lines. 28 redundant copies reduced to 1 each
within the event_log module.
