+++ crates/scp-core/src/event_log/proof.rs
@@ -13,8 +13,6 @@
 //!
 //! See ADR-011 in `.docs/adrs/phase-2.md` for the full design.
 
-use sha2::{Digest, Sha256};
-
 use super::{EventLog, EventLogError};
 use crate::event_log::tree;
 
@@ -324,14 +322,7 @@ pub fn verify_inclusion(proof: &InclusionProof) -> bool {
 // Internal helpers
 // ---------------------------------------------------------------------------
 
-/// Computes `SHA-256(0x01 || left || right)` for an interior node (RFC 6962 §2.1).
-fn hash_pair(left: &[u8; 32], right: &[u8; 32]) -> [u8; 32] {
-    let mut hasher = Sha256::new();
-    hasher.update([0x01]);
-    hasher.update(left);
-    hasher.update(right);
-    hasher.finalize().into()
-}
+use super::tree::hash_pair;
 
 // ---------------------------------------------------------------------------
 // Tests
@@ -340,145 +331,14 @@ fn hash_pair(left: &[u8; 32], right: &[u8; 32]) -> [u8; 32] {
 #[cfg(test)]
 #[allow(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
 mod tests {
-    use ed25519_dalek::Signer;
-    use sha2::{Digest, Sha256};
-
     use super::*;
-    use crate::event_log::tree::{self, GENESIS_PREV_HASH};
-    use crate::event_log::{Event, EventLog, EventLogError, EventPayload, EventType};
-
-    // -------------------------------------------------------------------
-    // Test helpers
-    // -------------------------------------------------------------------
-
-    /// Helper: create a signing keypair.
-    fn test_keypair() -> (ed25519_dalek::VerifyingKey, ed25519_dalek::SigningKey) {
-        let mut rng = rand::thread_rng();
-        let signing_key = ed25519_dalek::SigningKey::generate(&mut rng);
-        let verifying_key = signing_key.verifying_key();
-        (verifying_key, signing_key)
-    }
-
-    /// Helper: encode a public key as a test DID (`did:key:<hex>`).
-    fn did_from_pubkey(verifying_key: &ed25519_dalek::VerifyingKey) -> String {
-        let hex: String = verifying_key
-            .as_bytes()
-            .iter()
-            .fold(String::new(), |mut acc, b| {
-                use std::fmt::Write;
-                let _ = write!(acc, "{b:02x}");
-                acc
-            });
-        format!("did:key:{hex}")
-    }
-
-    /// Helper: compute the canonical hash for signing an event.
-    /// Must match the production `compute_event_canonical_hash` in `tree.rs`.
-    fn compute_event_canonical_hash(event: &Event) -> Vec<u8> {
-        let mut hasher = Sha256::new();
-        hasher.update(b"SCP-EVENT-V1:");
-        #[allow(clippy::cast_possible_truncation)]
-        let length_prefix = |hasher: &mut Sha256, bytes: &[u8]| {
-            hasher.update((bytes.len() as u32).to_be_bytes());
-            hasher.update(bytes);
-        };
-        hasher.update(event_type_tag(&event.event_type).to_be_bytes());
-        length_prefix(&mut hasher, event.actor_did.as_bytes());
-        hasher.update(event.timestamp.to_be_bytes());
-        hasher.update(event.sequence.to_be_bytes());
-        length_prefix(&mut hasher, &event.payload.data);
-        hasher.update(event.prev_hash);
-        hasher.finalize().to_vec()
-    }
-
-    /// Returns a stable numeric tag for each event type variant.
-    const fn event_type_tag(event_type: &EventType) -> u16 {
-        match event_type {
-            EventType::ContextCreated => 0,
-            EventType::ContextClosing => 1,
-            EventType::ContextClosed => 2,
-            EventType::ContextExpired => 3,
-            EventType::MemberJoined => 4,
-            EventType::MemberLeft => 5,
-            EventType::RoleAssigned => 6,
-            EventType::TokenRevoked => 7,
-            EventType::MessageSent => 8,
-            EventType::ToolRegistered => 9,
-            EventType::ToolUpdated => 10,
-            EventType::ToolInvoked => 11,
-            EventType::ToolVerified => 12,
-            EventType::ToolInterfaceEstablished => 13,
-            EventType::GovernanceAction => 14,
-            EventType::ConsistencyCheckpoint => 15,
-            EventType::AbsenceProofRequested => 16,
-            EventType::MemberBlocked => 17,
-            EventType::KeyEpochAdvance => 18,
-            EventType::MediaSessionStarted => 19,
-            EventType::MediaSessionEnded => 20,
-            EventType::PaymentReceived => 21,
-            EventType::EconomicPolicyChanged => 22,
-            EventType::SpendingUcanGranted => 23,
-            EventType::SpendingUcanRevoked => 24,
-        }
-    }
-
-    /// Helper: sign an event.
-    fn sign_event(
-        event_type: EventType,
-        actor_did: &str,
-        timestamp: u64,
-        sequence: u64,
-        payload: Vec<u8>,
-        prev_hash: [u8; 32],
-        signing_key: &ed25519_dalek::SigningKey,
-    ) -> Event {
-        let mut event = Event {
-            event_type,
-            actor_did: actor_did.into(),
-            timestamp,
-            sequence,
-            payload: EventPayload { data: payload },
-            prev_hash,
-            signature: Vec::new(),
-        };
-
-        let canonical_hash = compute_event_canonical_hash(&event);
-        let signature = signing_key.sign(&canonical_hash);
-        event.signature = signature.to_bytes().to_vec();
-
-        event
-    }
+    use crate::event_log::test_helpers::build_test_log;
+    use crate::event_log::tree;
+    use crate::event_log::{EventLog, EventLogError};
 
     /// Helper: build a log with `n` events and return the log and leaf hashes.
     fn build_log(n: u64) -> (EventLog, Vec<[u8; 32]>) {
-        let (verifying_key, signing_key) = test_keypair();
-        let did = did_from_pubkey(&verifying_key);
-        let mut log = EventLog::new("ctx-proof-test".to_owned());
-        let mut prev_hash = GENESIS_PREV_HASH;
-        let mut leaf_hashes = Vec::new();
-
-        for i in 0..n {
-            let event = sign_event(
-                EventType::MessageSent,
-                &did,
-                1_000_000 + i,
-                i,
-                format!("message {i}").into_bytes(),
-                prev_hash,
-                &signing_key,
-            );
-            tree::append(&mut log, &event).unwrap();
-            // Compute leaf hash with RFC 6962 domain separation (0x00 prefix).
-            let serialized = rmp_serde::to_vec(&event).unwrap();
-            let mut hasher = Sha256::new();
-            hasher.update([0x00]);
-            hasher.update(&serialized);
-            let leaf_hash: [u8; 32] = hasher.finalize().into();
-            leaf_hashes.push(leaf_hash);
-            prev_hash = leaf_hash;
-        }
-
-        (log, leaf_hashes)
+        build_test_log(n)
     }
 
     // -------------------------------------------------------------------
